Extract CSV worker parsing hook
